# sisged-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SISGED API",
        "description": "API REST do sistema SISGED. Documentação gerada com L5-Swagger (OpenAPI 3).",
        "contact": {
            "email": "contato@sisged.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor da API"
        }
    ],
    "paths": {
...
        "/cursos": {
            "get": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Lista todos os cursos",
                "operationId": "b96eabc1969edbe2bb0436759523b3a8",
                "responses": {
                    "200": {
                        "description": "Lista de cursos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Curso"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Cadastra um novo curso",
                "operationId": "42b1ddb43b31bbb9a34efa2b3139705f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Curso"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Curso criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Curso"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/aluno/cursos": {
            "get": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Lista somente o nome dos cursos (acesso do aluno)",
                "operationId": "019341c4603bc54e7613f2e59fb2d5cb",
                "responses": {
                    "200": {
                        "description": "Lista de nomes de cursos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cursos/{idCurso}": {
            "get": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Exibe um curso específico",
                "operationId": "8d01be01c463087da2d731e816cceb12",
                "parameters": [
                    {
                        "name": "idCurso",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Curso encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Curso"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Curso não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Atualiza um curso",
                "operationId": "92a837f25a7e8efdca390deee6b2e10f",
                "parameters": [
                    {
                        "name": "idCurso",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Curso"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Curso atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Curso"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Curso não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Remove um curso",
                "operationId": "a4ec09d080237d7b101fe0868b2807f1",
                "parameters": [
                    {
                        "name": "idCurso",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Curso removido"
                    },
                    "404": {
                        "description": "Curso não encontrado"
                    }
                }
            }
        },
...
        "/materias": {
            "get": {
                "tags": [
                    "Materias"
                ],
                "summary": "Lista todas as matérias",
                "operationId": "dc0546c43ffb8b5ba7e7022c1dd47669",
                "responses": {
                    "200": {
                        "description": "Lista de matérias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Materia"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Materias"
                ],
                "summary": "Cadastra uma nova matéria",
                "operationId": "25e80a7056e8d10469e998046d85909c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Materia"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Matéria criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Materia"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/materias/{idMateria}": {
            "get": {
                "tags": [
                    "Materias"
                ],
                "summary": "Exibe uma matéria específica",
                "operationId": "fedc96dd0a2e15e06efbdfee63ceea51",
                "parameters": [
                    {
                        "name": "idMateria",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Matéria encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Materia"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Matéria não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Materias"
                ],
                "summary": "Atualiza uma matéria",
                "operationId": "887ac10378c975a52aa1ace5fb6f29db",
                "parameters": [
                    {
                        "name": "idMateria",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Materia"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Matéria atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Materia"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Matéria não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Materias"
                ],
                "summary": "Remove uma matéria",
                "operationId": "1cedf95579641e806a3bdfa3ca953dbd",
                "parameters": [
                    {
                        "name": "idMateria",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Matéria removida"
                    },
                    "404": {
                        "description": "Matéria não encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Curso": {
                "title": "Curso",
                "required": [
                    "nomeCurso"
                ],
                "properties": {
                    "idCurso": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "Turma_idTurma": {
                        "type": "integer",
                        "example": 1
                    },
                    "nomeCurso": {
                        "type": "string",
                        "example": "Desenvolvimento Web"
                    },
                    "modalidadeCurso": {
                        "type": "string",
                        "example": "EAD"
                    },
                    "cargahorariaCurso": {
                        "type": "integer",
                        "example": 160
                    },
                    "nivelCurso": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "Materia": {
                "title": "Materia",
                "required": [
                    "nomeMateria"
                ],
                "properties": {
                    "idMateria": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "siglaMateria": {
                        "type": "string",
                        "example": "POO"
                    },
                    "nomeMateria": {
                        "type": "string",
                        "example": "Programação Orientada a Objetos"
                    },
                    "cargahorariaMateria": {
                        "type": "string",
                        "example": "40:00:00"
                    },
                    "ementaMateria": {
                        "type": "string",
                        "example": "Conceitos de POO em PHP e Java"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}